@@ -1,4 +1,4 @@
-// swift-tools-version:5.1
+// swift-tools-version:5.5
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "YipYipSwift",
     platforms: [
-        .iOS(.v10)
+        .iOS(.v15)
     ],
     products: [
         // Products define the executables and libraries produced by a package, and make them visible to other packages.
