@@ -23,7 +23,7 @@ internal struct SwiftUIMain: View {
     internal weak var delegate: SwiftUIMainDelegate?
     
     internal var body: some View {
-        NavigationView {
+        NavigationStack {
             ZStack {
                 GeometryReader { geometry in
                     VStack(spacing: 16) {
@@ -133,15 +133,15 @@ internal struct SwiftUIMain: View {
                     .background(Color.blue)
                     .edgesIgnoringSafeArea(.top)
                 }
-                
-                NavigationLink(
-                    destination: PagedTableViewViewControllerRepresentable(),
-                    isActive: self.$state.showPagedTableView
-                ) {
-                    EmptyView()
+            }
+            .navigationBarHidden(true)
+            .navigationDestination(
+                isPresented: self.$state.showPagedTableView,
+                destination: {
+                    PagedTableViewViewControllerRepresentable()
+                        .navigationBarTitleDisplayMode(.inline)
                 }
-                .navigationBarTitleDisplayMode(.inline)
-            }.navigationBarHidden(true)
+            )
         }
     }
     
